_requirements.txt
absl-py==0.7.1
asn1crypto==1.0.1
astor==0.8.0
atari-py==0.2.6
autopep8==1.4.4
backcall==0.1.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cloudpickle==1.2.1
cryptography==2.7
cycler==0.10.0
decorator==4.4.0
empyrical==0.5.3
entrypoints==0.3
flake8==3.7.8
future==0.17.1
gast==0.2.2
gdax-api==1.2.0.dev1
google-pasta==0.1.7
grpcio==1.22.0
gym==0.14.0
h5py==2.9.0
idna==2.8
inflection==0.3.1
ipykernel==5.1.1
ipython==7.7.0
ipython-genutils==0.2.0
jedi==0.14.1
joblib==0.13.2
jupyter-client==5.3.1
jupyter-core==4.5.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lxml==4.4.1
Markdown==3.1.1
matplotlib==3.1.1
mccabe==0.6.1
more-itertools==5.0.0
mpi4py==3.0.2
mpl-finance==0.10.0
ndg-httpsclient==0.5.1
numpy==1.17.0
opencv-python==4.1.1.26
pandas==0.25.0
pandas-datareader==0.7.4
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
pkg-resources==0.0.0
prompt-toolkit==2.0.9
protobuf==3.9.0
ptyprocess==0.6.0
pyasn1==0.4.7
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
pyfolio==0.9.2
pyglet==1.3.2
Pygments==2.4.2
pyOpenSSL==19.0.0
pyparsing==2.4.1.1
python-dateutil==2.8.0
pytz==2019.1
pyzmq==18.0.2
Quandl==3.4.8
requests==2.22.0
scikit-learn==0.21.2
scipy==1.3.0
seaborn==0.9.0
six==1.12.0
sklearn==0.0
stable-baselines==2.7.0
TA-Lib==0.4.17
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
tornado==6.0.3
traitlets==4.3.2
urllib3==1.25.3
wcwidth==0.1.7
Werkzeug==0.15.5
wrapt==1.11.2
